Modularize code more and load initial data from github
This is a bit of a mish-mash as I sorted out how to make the remote
calls to Github while still using the Octokit library.

I originally wanted to put it in a remote, and something about that
still /feels/ right, but for now I've just put the remote call in the
`action` side of the mutation and triggered them on app start.

Next up, loading repo data!