Switch to datascript as backing store
Asami had some NPE issues when handling joins that for things that
didn't exist -- f.e. a given value for an attribute doesn't exist in
the DB -- and I realized that fighting with the DB library isn't where
I should be spending my time, so I've switched to something that's
sufficient for the exercise.