Simplify track function
This is simpler in part because datascript is simpler than asami, and
simpler because instead of leaning on :db/id and friends, I'm using a
slightly dirtier approach of simply querying for the repo before doing
the rest of the logic.

I'm not sure if that's something I'd keep if this were a production
app, but it's not, so that doesn't matter.