Fix base-req-options initialization
`github-secret` isn't setup by mount when the namespace is originally
compiled, which means that anything which uses it only at compile time
will have the pre-started version, and thus not work.

I've resolved this by changing `base-rep-options` to a function so
that it will get the up-to-date version when it's called.